pyproject.toml
[tool.poetry]
name = "pyanide"
version = "1.0.0"
description = "Tools for killing processes and keeping them dead."
authors = ["Matthew Martin <matthewdeanmartin@gmail.com>"]
keywords = ["kill", "exit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
]
include = [
    "pyanide/**/*.py",
    "pyanide/**/*.md",
    "pyanide/**/*.txt",
    "pyanide/**/*.html",
    "pyanide/**/*.jinja",
]
exclude = [
]
license = "MIT"
readme = "README.md"
repository = "https://github.com/matthewdeanmartin/pyanide"
homepage = "https://github.com/matthewdeanmartin/pyanide"
documentation ="https://github.com/matthewdeanmartin/pyanide"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/matthewdeanmartin/pyanide/issues"
"Change Log" = "https://github.com/matthewdeanmartin/pyanide/blob/main/CHANGES.md"

[tool.poetry.dependencies]
# DOM parser
mistune = "*"
# alternate formatter
mdformat = "*"
# support more types to table
tabulate = "*"


[tool.poetry.dev-dependencies]
#marko = "*"
#mistletoe = "*"
pytest = "*"
vermin = "*"
hypothesis = {extras = ["cli"], version = "*"}
pre-commit = "*"
mypy = "*"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
junit_family = "xunit1"
norecursedirs = ["vendor", "scripts"]
# don't know how to do this in toml
#addopts = "--strict-markers"
#markers =
#	slow: marks tests as slow (deselect with '-m "not slow"')
#	fast: marks tests as fast (deselect with '-m "not fast"')

[tool.isort]
default_section = "THIRDPARTY"
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = ["pyanide"]
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.ruff]
line-length = 1000

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = [
    "E722",
    "E501",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    "dead_code",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "hide"
]
per-file-ignores = { }

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py311"

[tool.scriv]
format = "md"
version = "literal: pyproject.toml: project.version"

[tool.hatch.version]
path = "pyanide/__init__.py"
output_file = "CHANGELOG.${config:format}"

#[tool.setuptools]
## find = {}
#py-modules = ["pyanide",
#    "pyanide.deserialize",
#    "pyanide.python_to_tables",
#    "pyanide.serialize",
#]